test: extract shared fakeDatastore into fake_datastore_test.go
Pure refactor — moves the fakeDatastore struct, all method impls, milpacs
panic stubs, and withTicketsKey helper from tickets_test.go into a new
shared file so the upcoming milpacs_test.go can reuse the same scaffolding.
Zero behavior change; all existing tests pass.
