+++ servers/grpc/tickets_test.go
@@ -1,7 +1,6 @@
 package grpc
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"testing"
@@ -9,7 +8,6 @@ import (
 	"github.com/7cav/api/datastores"
 	"github.com/7cav/api/proto"
 	"github.com/7cav/api/referencecache"
-	"github.com/7cav/api/xenforo"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc/codes"
@@ -17,63 +15,6 @@ import (
 	"google.golang.org/protobuf/types/known/emptypb"
 	"gorm.io/gorm"
 )
-
-// fakeDatastore implements datastores.Datastore with just the methods we need
-// for tickets-handler tests stubbed. Other methods panic so the test surface
-// stays explicit.
-type fakeDatastore struct {
-	listTickets func(*datastores.ListTicketsFilter) ([]*proto.Ticket, string, bool, error)
-	getTicket   func(uint32) (*proto.Ticket, error)
-	getByRef    func(string) (*proto.Ticket, error)
-	firstMsgs   func(uint32, int) ([]*proto.Message, uint32, error)
-	listMsgs    func(uint32, string, uint32, bool) ([]*proto.Message, string, bool, error)
-	listCats    func() ([]*proto.Category, error)
-}
-
-func (f *fakeDatastore) ListTickets(_ context.Context, _ datastores.TicketReferenceCache, fi *datastores.ListTicketsFilter) ([]*proto.Ticket, string, bool, error) {
-	return f.listTickets(fi)
-}
-func (f *fakeDatastore) GetTicket(_ context.Context, _ datastores.TicketReferenceCache, id uint32, _ string) (*proto.Ticket, error) {
-	return f.getTicket(id)
-}
-func (f *fakeDatastore) GetTicketByRef(_ context.Context, _ datastores.TicketReferenceCache, ref string, _ string) (*proto.Ticket, error) {
-	return f.getByRef(ref)
-}
-func (f *fakeDatastore) GetTicketFirstMessages(_ context.Context, id uint32, n int, _ bool) ([]*proto.Message, uint32, error) {
-	return f.firstMsgs(id, n)
-}
-func (f *fakeDatastore) ListTicketMessages(_ context.Context, id uint32, after string, per uint32, hidden bool) ([]*proto.Message, string, bool, error) {
-	return f.listMsgs(id, after, per, hidden)
-}
-func (f *fakeDatastore) ListCategories(_ context.Context, _ datastores.TicketReferenceCache) ([]*proto.Category, error) {
-	return f.listCats()
-}
-
-// Milpacs methods omitted; tickets tests don't use them.
-func (f *fakeDatastore) FindProfilesById(_ ...uint64) ([]*proto.Profile, error)           { panic("unused") }
-func (f *fakeDatastore) FindProfilesByUsername(string) ([]*proto.Profile, error)          { panic("unused") }
-func (f *fakeDatastore) FindRosterByType(proto.RosterType) (*proto.Roster, error)         { panic("unused") }
-func (f *fakeDatastore) FindLiteRosterByType(proto.RosterType) (*proto.LiteRoster, error) { panic("unused") }
-func (f *fakeDatastore) FindProfileByKeycloakID(string) (*proto.Profile, error)           { panic("unused") }
-func (f *fakeDatastore) FindProfileByDiscordID(string) (*proto.Profile, error)            { panic("unused") }
-func (f *fakeDatastore) FindProfilesByPosition(string) (*proto.LiteRoster, error)         { panic("unused") }
-func (f *fakeDatastore) FindS1UniformsRosterByType(proto.RosterType) (*proto.S1UniformsRoster, error) {
-	panic("unused")
-}
-func (f *fakeDatastore) FindAllRanks() ([]*proto.RankExpanded, error)             { panic("unused") }
-func (f *fakeDatastore) FindAllPositionGroups() ([]*proto.PositionGroup, error)   { panic("unused") }
-func (f *fakeDatastore) FindAwol() ([]*proto.Awol, error)                         { panic("unused") }
-func (f *fakeDatastore) GetTableUpdates() ([]xenforo.TableInfo, error)            { panic("unused") }
-func (f *fakeDatastore) FindProfileByGamertag(string) (*proto.Profile, error)     { panic("unused") }
-func (f *fakeDatastore) ValidateApiKey(string) (*datastores.ApiKeyResult, error)  { panic("unused") }
-
-func withTicketsKey(scopes ...string) context.Context {
-	m := map[string]struct{}{}
-	for _, s := range scopes {
-		m[s] = struct{}{}
-	}
-	return ContextWithKey(context.Background(), &datastores.ApiKeyResult{Scopes: m})
-}
 
 func TestListTickets_RequiresScope(t *testing.T) {
 	svc := &TicketsService{Datastore: &fakeDatastore{}, ReferenceCache: &referencecache.Cache{}}
